refactor: remove TeamDetails component and integrate its functionality into TeamMetaSection
