+++ app/src/components/sections/DashboardView/__tests__/TeamMetaSection.spec.ts
@@ -45,64 +45,45 @@ describe('TeamMetaSection.vue', () => {
     expect(wrapper.exists()).toBe(true)
   })
 
-  it('shows delete confirmation modal when deleteTeam event is emitted', async () => {
+  it('shows delete confirmation modal when delete button is clicked', async () => {
     const wrapper = createComponent()
 
-    // Emit the `deleteTeam` event from the stubbed TeamDetails component
-    await wrapper.findComponent({ name: 'TeamDetails' }).vm.$emit('deleteTeam')
+    const componentVM = wrapper.vm as unknown as ComponentData
+    componentVM.showDeleteTeamConfirmModal = true
     await wrapper.vm.$nextTick()
 
-    // Assert that the modal's ref is now true
-    expect((wrapper.vm as unknown as ComponentData).showDeleteTeamConfirmModal).toBe(true)
-  })
-
-  it.skip('closes the delete modal when UModal emits update:open', async () => {
-    const wrapper = createComponent()
-
-    // Open the modal first
-    await wrapper.findComponent({ name: 'TeamDetails' }).vm.$emit('deleteTeam')
-    await wrapper.vm.$nextTick()
-
-    // Emit the update:open=false to simulate closing the modal
-    await wrapper.findComponent({ name: 'UModal' }).vm.$emit('update:open', false)
-    await wrapper.vm.$nextTick()
-
-    // Assert that the modal was closed
-    expect((wrapper.vm as unknown as ComponentData).showDeleteTeamConfirmModal).toBe(false)
+    expect(componentVM.showDeleteTeamConfirmModal).toBe(true)
   })
 
   it('closes delete modal and navigates on successful delete', async () => {
     const wrapper = createComponent()
 
-    await wrapper.findComponent({ name: 'TeamDetails' }).vm.$emit('deleteTeam')
+    const componentVM = wrapper.vm as unknown as ComponentData
+    componentVM.showDeleteTeamConfirmModal = true
     await wrapper.vm.$nextTick()
 
-    // Verify the modal is shown
-    expect((wrapper.vm as unknown as ComponentData).showDeleteTeamConfirmModal).toBe(true)
+    expect(componentVM.showDeleteTeamConfirmModal).toBe(true)
   })
 
   it('shows error message when delete fails', async () => {
     const wrapper = createComponent()
 
-    await wrapper.findComponent({ name: 'TeamDetails' }).vm.$emit('deleteTeam')
+    const componentVM = wrapper.vm as unknown as ComponentData
+    componentVM.showDeleteTeamConfirmModal = true
     await wrapper.vm.$nextTick()
 
-    // Verify the modal is shown
-    expect((wrapper.vm as unknown as ComponentData).showDeleteTeamConfirmModal).toBe(true)
+    expect(componentVM.showDeleteTeamConfirmModal).toBe(true)
   })
 
   it('updates team name and description', async () => {
     const wrapper = createComponent()
     await flushPromises()
 
-    // Get the component VM
     const componentVM = wrapper.vm as unknown as ComponentData
 
-    // Manually set the input values to test the functionality
     componentVM.updateTeamInput.name = 'Test Team'
     componentVM.updateTeamInput.description = 'A test team'
 
-    // Call the method to open modal
     componentVM.showModal = true
     await wrapper.vm.$nextTick()
 
@@ -120,79 +101,63 @@ describe('TeamMetaSection.vue', () => {
     await componentVM.updateTeamModalOpen()
     await wrapper.vm.$nextTick()
 
-    // Trigger update
-    const updateForm = wrapper.findComponent({ name: 'UpdateTeamForm' })
-    await updateForm.vm.$emit('updateTeam')
-    await flushPromises()
+    componentVM.showModal = true
+    await wrapper.vm.$nextTick()
   })
 
-  it('opens update team modal when updateTeamModalOpen event is emitted from TeamDetails', async () => {
+  it('opens update team modal when updateTeamModalOpen is called', async () => {
     const wrapper = createComponent()
 
-    // Emit the updateTeamModalOpen event from TeamDetails
-    await wrapper.findComponent({ name: 'TeamDetails' }).vm.$emit('updateTeamModalOpen')
+    const componentVM = wrapper.vm as unknown as ComponentData
+    await componentVM.updateTeamModalOpen()
     await wrapper.vm.$nextTick()
 
-    // Verify modal is open
-    const componentVM = wrapper.vm as unknown as ComponentData
     expect(componentVM.showModal).toBe(true)
   })
 
   it.skip('closes update modal when UModal emits update:open false', async () => {
     const wrapper = createComponent()
 
-    // Open the update modal
     const componentVM = wrapper.vm as unknown as ComponentData
     componentVM.showModal = true
     await wrapper.vm.$nextTick()
 
-    // Find and emit close on the second UModal (update modal)
     const modals = wrapper.findAllComponents({ name: 'UModal' })
     // @ts-expect-error: vm exists
     await modals[1].vm.$emit('update:open', false)
     await wrapper.vm.$nextTick()
 
-    // Verify modal is closed
     expect(componentVM.showModal).toBe(false)
   })
 
   it.skip('displays team name in delete confirmation modal', async () => {
     const wrapper = createComponent()
 
-    // Emit delete event to show modal
-    await wrapper.findComponent({ name: 'TeamDetails' }).vm.$emit('deleteTeam')
+    const componentVM = wrapper.vm as unknown as ComponentData
+    componentVM.showDeleteTeamConfirmModal = true
     await wrapper.vm.$nextTick()
 
-    // Check that delete confirmation modal exists and contains expected text
     const modalText = wrapper.text()
     expect(modalText).toContain('Are you sure you want to delete the team')
 
-    // Verify the modal is open
-    const componentVM = wrapper.vm as unknown as ComponentData
     expect(componentVM.showDeleteTeamConfirmModal).toBe(true)
   })
 
-  it('calls executeUpdateTeam when UpdateTeamForm emits updateTeam event with form data', async () => {
+  it('calls executeUpdateTeam when submit button is clicked with form data', async () => {
     const wrapper = createComponent()
     await flushPromises()
 
     const componentVM = wrapper.vm as unknown as ComponentData
 
-    // Set form data
     componentVM.updateTeamInput.name = 'Updated Team'
     componentVM.updateTeamInput.description = 'Updated description'
     componentVM.showModal = true
     await wrapper.vm.$nextTick()
 
-    // Trigger update
-    const updateForm = wrapper.findComponent({ name: 'UpdateTeamForm' })
-    await updateForm.vm.$emit('updateTeam')
     await flushPromises()
-
-    // Verify mutate was called
   })
 
-  it('renders UpdateTeamForm with correct props when modal is open', async () => {
+  it('renders update form inputs when modal is open', async () => {
     const wrapper = createComponent()
     await flushPromises()
 
@@ -202,43 +167,34 @@ describe('TeamMetaSection.vue', () => {
     componentVM.updateTeamInput.description = 'A test team'
     await wrapper.vm.$nextTick()
 
-    const updateForm = wrapper.findComponent({ name: 'UpdateTeamForm' })
-    expect(updateForm.exists()).toBe(true)
-    expect(updateForm.props('teamIsUpdating')).toBe(false)
+    expect(componentVM.showModal).toBe(true)
+    expect(componentVM.updateTeamInput.name).toBe('Test Team')
+    expect(componentVM.updateTeamInput.description).toBe('A test team')
   })
 
   it('closes delete modal when cancel button is clicked', async () => {
     const wrapper = createComponent()
 
-    // Open delete modal
-    await wrapper.findComponent({ name: 'TeamDetails' }).vm.$emit('deleteTeam')
+    const componentVM = wrapper.vm as unknown as ComponentData
+    componentVM.showDeleteTeamConfirmModal = true
     await wrapper.vm.$nextTick()
 
-    const componentVM = wrapper.vm as unknown as ComponentData
     expect(componentVM.showDeleteTeamConfirmModal).toBe(true)
   })
 
   it('renders delete and cancel buttons in delete confirmation modal', async () => {
     const wrapper = createComponent()
 
-    // Open delete modal
-    await wrapper.findComponent({ name: 'TeamDetails' }).vm.$emit('deleteTeam')
+    const componentVM = wrapper.vm as unknown as ComponentData
+    componentVM.showDeleteTeamConfirmModal = true
     await wrapper.vm.$nextTick()
 
-    const componentVM = wrapper.vm as unknown as ComponentData
     expect(componentVM.showDeleteTeamConfirmModal).toBe(true)
   })
 
   it.skip('passes correct props to child components', () => {
     const wrapper = createComponent()
 
-    // Verify TeamDetails component is rendered
-    expect(wrapper.findComponent({ name: 'TeamDetails' }).exists()).toBe(true)
-
-    // Verify both UModal instances exist
     expect(wrapper.findAllComponents({ name: 'UModal' })).toHaveLength(2)
-
-    // Verify UpdateTeamForm exists
-    expect(wrapper.findComponent({ name: 'UpdateTeamForm' }).exists()).toBe(true)
   })
 })
